Small 64-bit fixes.
git-svn-id: https://svn.tdr.iem.uni-due.de:8443/svn/sctp/trunk/rsplib2@1335 516812f1-d3fd-0310-9ea4-9aededebcb87

+++ rsplib/componentstatusreporter.c
@@ -106,7 +106,7 @@ void cspReporterNew(struct CSPReporter*    cspReporter,
                     const unsigned int     cspReportInterval,
                     size_t                 (*cspGetReportFunction)(
                                               void*                         userData,
-                                              unsigned long long*           identifier,
+                                              uint64_t*                     identifier,
                                               struct ComponentAssociation** caeArray,
                                               char*                         statusText,
                                               char*                         componentAddress,
