+++ rsplib/componentstatusreporter.h
@@ -59,7 +59,7 @@ struct CSPReporter
 
    size_t               (*CSPGetReportFunction)(
                            void*                         userData,
-                           unsigned long long*           identifier,
+                           uint64_t*                     identifier,
                            struct ComponentAssociation** caeArray,
                            char*                         statusText,
                            char*                         componentAddress,
@@ -75,7 +75,7 @@ void cspReporterNew(struct CSPReporter*    cspReporter,
                     const unsigned int     cspReportInterval,
                     size_t                 (*cspGetReportFunction)(
                                               void*                         userData,
-                                              unsigned long long*           identifier,
+                                              uint64_t*                     identifier,
                                               struct ComponentAssociation** caeArray,
                                               char*                         statusText,
                                               char*                         componentAddress,
